+++ src/Asset/Elements/Asset.php
@@ -108,6 +108,7 @@ use Stringable;
 use Throwable;
 use Twig\Markup;
 
+use function CraftCms\Cms\currentUser;
 use function CraftCms\Cms\currentUserElement;
 use function CraftCms\Cms\t;
 
@@ -1534,7 +1535,7 @@ JS, [
 
         if (
             app(ElementRequest::class)->element() === $this &&
-            Auth::craftUser()->isAdmin() &&
+            currentUser()->isAdmin() &&
             Cms::config()->allowAdminChanges
         ) {
             $items[] = ['type' => MenuItemType::HR];
@@ -2519,7 +2520,7 @@ JS, [
         // See if we can show a thumbnail
         try {
             // Is the image editable, and is the user allowed to edit?
-            $user = Auth::craftUser();
+            $user = currentUser();
             $previewable = AssetsService::getAssetPreviewHandler($this) !== null;
             $editable = (
                 $this->getSupportsImageEditor() &&
@@ -3057,7 +3058,7 @@ JS;
                 ],
             ];
 
-            $user = Auth::craftUser();
+            $user = currentUser();
 
             if ($user && Gate::check('moveFolderFrom', $this->getFolder())) {
                 $attributes['data']['movable'] = true;